+++ tests/unit/test_main.py
@@ -1,6 +1,6 @@
 """Tests for the main madule."""
 
-from unittest.mock import patch, MagicMock
+from unittest.mock import patch, call, MagicMock
 from datetime import datetime, timedelta
 
 import pytz
@@ -22,10 +22,12 @@ class TestMain:
     def setup_method(self):
         """Initialize before tests are executed."""
         self.controller = get_controller_mock()
-        self.controller.napps[('kytos', 'topology')] = MagicMock()
+        self.controller.napps[("kytos", "topology")] = MagicMock()
         self.controller.switches = MagicMock()
         self.scheduler = MagicMock()
-        new_sched = 'napps.kytos.maintenance.managers.MaintenanceScheduler.new_scheduler'
+        new_sched = (
+            "napps.kytos.maintenance.managers.MaintenanceScheduler.new_scheduler"
+        )
         with patch(new_sched) as new_scheduler:
             new_scheduler.return_value = self.scheduler
             self.napp = Main(self.controller)
@@ -36,42 +38,183 @@ class TestMain:
     async def test_create_mw_case_1(self, event_loop):
         """Test a successful case of the REST to create."""
         self.napp.controller.loop = event_loop
-        url = f'{self.base_endpoint}'
+        url = f"{self.base_endpoint}"
         start = datetime.now(pytz.utc) + timedelta(days=1)
         end = start + timedelta(hours=2)
+        self.controller.switches = MagicMock()
+        self.controller.get_interface_by_id = MagicMock()
+        self.controller.napps[("kytos", "topology")].links = MagicMock()
+        switches = {"00:00:00:00:00:00:00:02": 1, "00:00:00:00:00:00:00:03": 2}
+        interfaces = {"00:00:00:00:00:00:00:03:3": 1, "00:00:00:00:00:00:00:02:1": 2}
+        links = {
+            "cf0f4071be426b3f745027f5d22bc61f8312ae86293c9b28e7e66015607a9260": 1,
+            "4d42dc0852278accac7d9df15418f6d921db160b13d674029a87cef1b5f67f30": 2,
+        }
+        self.controller.switches.get.side_effect = switches.get
+        get_interface_side_effect = lambda interface_id: interfaces.get(interface_id)
+        self.controller.get_interface_by_id.side_effect = get_interface_side_effect
+        self.controller.napps[("kytos", "topology")].links.get.side_effect = links.get
         payload = {
             "start": start.strftime(TIME_FMT),
             "end": end.strftime(TIME_FMT),
-            "switches": [
-                "00:00:00:00:00:00:02",
-            ],
-            'interfaces': [
-                "00:00:00:00:00:00:00:03:3",
+            "switches": ["00:00:00:00:00:00:00:02", "00:00:00:00:00:00:00:03"],
+            "interfaces": ["00:00:00:00:00:00:00:03:3", "00:00:00:00:00:00:00:02:1"],
+            "links": [
+                "cf0f4071be426b3f745027f5d22bc61f8312ae86293c9b28e7e66015607a9260",
+                "4d42dc0852278accac7d9df15418f6d921db160b13d674029a87cef1b5f67f30",
             ],
         }
         response = await self.api.post(url, json=payload)
         current_data = response.json()
         assert response.status_code == 201, current_data
+        switch_calls = [
+            call("00:00:00:00:00:00:00:02"),
+            call("00:00:00:00:00:00:00:03"),
+        ]
+        self.controller.switches.get.assert_has_calls(switch_calls)
+        interface_calls = [
+            call("00:00:00:00:00:00:00:03:3"),
+            call("00:00:00:00:00:00:00:02:1"),
+        ]
+        self.controller.get_interface_by_id.assert_has_calls(interface_calls)
+        link_calls = [
+            call("cf0f4071be426b3f745027f5d22bc61f8312ae86293c9b28e7e66015607a9260"),
+            call("4d42dc0852278accac7d9df15418f6d921db160b13d674029a87cef1b5f67f30"),
+        ]
+        self.controller.napps[("kytos", "topology")].links.get.assert_has_calls(
+            link_calls
+        )
         args, kwargs = self.scheduler.add.call_args
         window: MW = args[0]
-    
+
         assert window.start == start.replace(microsecond=0)
         assert window.end == end.replace(microsecond=0)
-        assert window.switches == ['00:00:00:00:00:00:02']
-        assert window.interfaces == ['00:00:00:00:00:00:00:03:3']
-        assert window.links == []
-        
-        assert current_data == {'mw_id': window.id}
+        assert window.switches == ["00:00:00:00:00:00:00:02", "00:00:00:00:00:00:00:03"]
+        assert window.interfaces == [
+            "00:00:00:00:00:00:00:03:3",
+            "00:00:00:00:00:00:00:02:1",
+        ]
+        assert window.links == [
+            "cf0f4071be426b3f745027f5d22bc61f8312ae86293c9b28e7e66015607a9260",
+            "4d42dc0852278accac7d9df15418f6d921db160b13d674029a87cef1b5f67f30",
+        ]
+
+        assert current_data == {"mw_id": window.id}
 
     async def test_create_mw_case_2(self, event_loop):
+        """Test a successful case of the REST to create."""
+        self.napp.controller.loop = event_loop
+        url = f"{self.base_endpoint}"
+        start = datetime.now(pytz.utc) + timedelta(days=1)
+        end = start + timedelta(hours=2)
+        self.controller.switches = MagicMock()
+        switches = {"00:00:00:00:00:00:00:01": 1}
+        self.controller.switches.get.side_effect = switches.get
+        payload = {
+            "start": start.strftime(TIME_FMT),
+            "end": end.strftime(TIME_FMT),
+            "switches": [
+                "00:00:00:00:00:00:00:01",
+            ],
+        }
+        response = await self.api.post(url, json=payload)
+        current_data = response.json()
+        assert response.status_code == 201, current_data
+        self.controller.switches.get.assert_called_with("00:00:00:00:00:00:00:01")
+        args, kwargs = self.scheduler.add.call_args
+        window: MW = args[0]
+
+        assert window.start == start.replace(microsecond=0)
+        assert window.end == end.replace(microsecond=0)
+        assert window.switches == ["00:00:00:00:00:00:00:01"]
+        assert window.interfaces == []
+        assert window.links == []
+
+        assert current_data == {"mw_id": window.id}
+
+    async def test_create_mw_case_3(self, event_loop):
+        """Test a successful case of the REST to create."""
+        self.napp.controller.loop = event_loop
+        url = f"{self.base_endpoint}"
+        start = datetime.now(pytz.utc) + timedelta(days=1)
+        end = start + timedelta(hours=2)
+        self.controller.get_interface_by_id = MagicMock()
+        interfaces = {"00:00:00:00:00:00:00:01:1": 1}
+        get_interface_side_effect = lambda interface_id: interfaces.get(interface_id)
+        self.controller.get_interface_by_id.side_effect = get_interface_side_effect
+        payload = {
+            "id": "some_mw",
+            "description": "some_mw",
+            "start": start.strftime(TIME_FMT),
+            "end": end.strftime(TIME_FMT),
+            "interfaces": [
+                "00:00:00:00:00:00:00:01:1",
+            ],
+        }
+        response = await self.api.post(url, json=payload)
+        current_data = response.json()
+        assert response.status_code == 201, current_data
+        self.controller.get_interface_by_id.assert_called_with(
+            "00:00:00:00:00:00:00:01:1"
+        )
+        assert response.status_code == 201, current_data
+        args, kwargs = self.scheduler.add.call_args
+        response = await self.api.get(f"{url}/some_mw")
+        window: MW = args[0]
+
+        assert window.start == start.replace(microsecond=0)
+        assert window.end == end.replace(microsecond=0)
+        assert window.switches == []
+        assert window.interfaces == ["00:00:00:00:00:00:00:01:1"]
+        assert window.links == []
+
+        assert current_data == {"mw_id": "some_mw"}
+
+    async def test_create_mw_case_4(self, event_loop):
+        """Test a successful case of the REST to create."""
+        self.napp.controller.loop = event_loop
+        url = f"{self.base_endpoint}"
+        start = datetime.now(pytz.utc) + timedelta(days=1)
+        end = start + timedelta(hours=2)
+        self.controller.napps[("kytos", "topology")].links = MagicMock()
+        links = {"cf0f4071be426b3f745027f5d22bc61f8312ae86293c9b28e7e66015607a9260": 1}
+        self.controller.napps[("kytos", "topology")].links.get.side_effect = links.get
+        payload = {
+            "start": start.strftime(TIME_FMT),
+            "end": end.strftime(TIME_FMT),
+            "links": [
+                "cf0f4071be426b3f745027f5d22bc61f8312ae86293c9b28e7e66015607a9260",
+            ],
+        }
+        response = await self.api.post(url, json=payload)
+        current_data = response.json()
+        assert response.status_code == 201, current_data
+        self.controller.napps[("kytos", "topology")].links.get.assert_called_with(
+            "cf0f4071be426b3f745027f5d22bc61f8312ae86293c9b28e7e66015607a9260"
+        )
+        assert response.status_code == 201, current_data
+        args, kwargs = self.scheduler.add.call_args
+        window: MW = args[0]
+
+        assert window.start == start.replace(microsecond=0)
+        assert window.end == end.replace(microsecond=0)
+        assert window.switches == []
+        assert window.interfaces == []
+        assert window.links == [
+            "cf0f4071be426b3f745027f5d22bc61f8312ae86293c9b28e7e66015607a9260"
+        ]
+
+        assert current_data == {"mw_id": window.id}
+
+    async def test_create_mw_case_5(self, event_loop):
         """Test a fail case of the REST to create a maintenance window."""
         self.napp.controller.loop = event_loop
-        url = f'{self.base_endpoint}'
+        url = f"{self.base_endpoint}"
         payload = {
             "switches": [
                 "00:00:00:00:00:00:02",
             ],
-            'interfaces': [
+            "interfaces": [
                 "00:00:00:00:00:00:00:03:3",
             ],
         }
@@ -79,10 +222,10 @@ class TestMain:
         assert response.status_code == 400
         self.scheduler.add.assert_not_called()
 
-    async def test_create_mw_case_3(self, event_loop):
+    async def test_create_mw_case_6(self, event_loop):
         """Test a fail case of the REST to create a maintenance window."""
         self.napp.controller.loop = event_loop
-        url = f'{self.base_endpoint}'
+        url = f"{self.base_endpoint}"
         start = datetime.now(pytz.utc) - timedelta(days=1)
         end = start + timedelta(hours=2)
         payload = {
@@ -91,21 +234,20 @@ class TestMain:
             "switches": [
                 "00:00:00:00:00:00:02",
             ],
-            'interfaces': [
+            "interfaces": [
                 "00:00:00:00:00:00:00:03:3",
             ],
         }
         response = await self.api.post(url, json=payload)
         current_data = response.json()
         assert response.status_code == 400
-        assert current_data['description'] == \
-                         'start: Start in the past not allowed'
+        assert current_data["description"] == "start: Start in the past not allowed"
         self.scheduler.add.assert_not_called()
 
-    async def test_create_mw_case_4(self, event_loop):
+    async def test_create_mw_case_7(self, event_loop):
         """Test a fail case of the REST to create a maintenance window."""
         self.napp.controller.loop = event_loop
-        url = f'{self.base_endpoint}'
+        url = f"{self.base_endpoint}"
         start = datetime.now(pytz.utc) + timedelta(days=1)
         end = start - timedelta(hours=2)
         payload = {
@@ -114,7 +256,7 @@ class TestMain:
             "switches": [
                 "00:00:00:00:00:00:02",
             ],
-            'interfaces': [
+            "interfaces": [
                 "00:00:00:00:00:00:00:03:3",
             ],
         }
@@ -122,25 +264,24 @@ class TestMain:
         current_data = response.json()
 
         assert response.status_code == 400
-        assert current_data['description'] == \
-                         'end: End before start not allowed'
+        assert current_data["description"] == "end: End before start not allowed"
         self.scheduler.add.assert_not_called()
 
     @pytest.mark.skip(reason="Future feature")
-    async def test_create_mw_case_5(self, event_loop):
+    async def test_create_mw_case_8(self, event_loop):
         """Test a fail case of the REST to create a maintenance window."""
         self.napp.controller.loop = event_loop
-        url = f'{self.base_endpoint}'
+        url = f"{self.base_endpoint}"
         start = datetime.now(pytz.utc) + timedelta(days=1)
         end = start + timedelta(hours=2)
         payload = {
-            'id': '1234',
+            "id": "1234",
             "start": start.strftime(TIME_FMT),
             "end": end.strftime(TIME_FMT),
             "switches": [
                 "00:00:00:00:00:00:02",
             ],
-            'interfaces': [
+            "interfaces": [
                 "00:00:00:00:00:00:00:03:3",
             ],
         }
@@ -148,24 +289,23 @@ class TestMain:
         current_data = response.json()
 
         assert response.status_code == 400
-        assert current_data['description'] == \
-                         'Setting a maintenance id is not allowed'
+        assert current_data["description"] == "Setting a maintenance id is not allowed"
         self.scheduler.add.assert_not_called()
 
-    async def test_create_mw_case_6(self, event_loop):
+    async def test_create_mw_case_9(self, event_loop):
         """Test a fail case of the REST to create a maintenance window."""
         self.napp.controller.loop = event_loop
-        url = f'{self.base_endpoint}'
+        url = f"{self.base_endpoint}"
         start = datetime.now(pytz.utc) + timedelta(days=1)
         end = start + timedelta(hours=2)
         payload = {
-            'status': 'fun',
+            "status": "fun",
             "start": start.strftime(TIME_FMT),
             "end": end.strftime(TIME_FMT),
             "switches": [
                 "00:00:00:00:00:00:02",
             ],
-            'interfaces': [
+            "interfaces": [
                 "00:00:00:00:00:00:00:03:3",
             ],
         }
@@ -173,14 +313,17 @@ class TestMain:
         current_data = response.json()
 
         assert response.status_code == 400
-        assert current_data['description'] == \
-                         'Setting a maintenance status is not allowed'
+        assert (
+            current_data["description"] == "Setting a maintenance status is not allowed"
+        )
         self.scheduler.add.assert_not_called()
 
     async def test_get_mw_case_1(self):
         """Test get all maintenance windows, empty list."""
-        self.scheduler.list_maintenances.return_value = MaintenanceWindows.construct(__root__ = [])
-        url = f'{self.base_endpoint}'
+        self.scheduler.list_maintenances.return_value = MaintenanceWindows.construct(
+            __root__=[]
+        )
+        url = f"{self.base_endpoint}"
         response = await self.api.get(url)
         current_data = response.json()
         assert response.status_code == 200
@@ -195,68 +338,61 @@ class TestMain:
         start2 = datetime.now(pytz.utc) + timedelta(hours=5)
         end2 = start2 + timedelta(hours=1, minutes=30)
         self.scheduler.list_maintenances.return_value = MaintenanceWindows.construct(
-            __root__ = [
-            MW.construct(
-                id = '1234',
-                start = start1.replace(microsecond=0),
-                end = end1.replace(microsecond=0),
-                switches = [
-                    '00:00:00:00:00:00:12:23'
-                ],
-                description = '',
-                links = [],
-                interfaces = [],
-                status = 'pending',
-                updated_at = now.replace(microsecond=0),
-                inserted_at = now.replace(microsecond=0),
-            ),
-            MW.construct(
-                id = '4567',
-                start = start2.replace(microsecond=0),
-                end = end2.replace(microsecond=0),
-                switches = [
-                    '12:34:56:78:90:ab:cd:ef'
-                ],
-                description = '',
-                links = [],
-                interfaces = [],
-                status = 'pending',
-                updated_at = now.replace(microsecond=0),
-                inserted_at = now.replace(microsecond=0),
-            ),
-        ])
+            __root__=[
+                MW.construct(
+                    id="1234",
+                    start=start1.replace(microsecond=0),
+                    end=end1.replace(microsecond=0),
+                    switches=["00:00:00:00:00:00:12:23"],
+                    description="",
+                    links=[],
+                    interfaces=[],
+                    status="pending",
+                    updated_at=now.replace(microsecond=0),
+                    inserted_at=now.replace(microsecond=0),
+                ),
+                MW.construct(
+                    id="4567",
+                    start=start2.replace(microsecond=0),
+                    end=end2.replace(microsecond=0),
+                    switches=["12:34:56:78:90:ab:cd:ef"],
+                    description="",
+                    links=[],
+                    interfaces=[],
+                    status="pending",
+                    updated_at=now.replace(microsecond=0),
+                    inserted_at=now.replace(microsecond=0),
+                ),
+            ]
+        )
         mw_dict = [
             {
-                'id': '1234',
-                'start': start1.strftime(TIME_FMT),
-                'end': end1.strftime(TIME_FMT),
-                'switches': [
-                    '00:00:00:00:00:00:12:23'
-                ],
-                'description': '',
-                'links': [],
-                'interfaces': [],
-                'status': 'pending',
-                'updated_at': now.strftime(TIME_FMT),
-                'inserted_at': now.strftime(TIME_FMT),
+                "id": "1234",
+                "start": start1.strftime(TIME_FMT),
+                "end": end1.strftime(TIME_FMT),
+                "switches": ["00:00:00:00:00:00:12:23"],
+                "description": "",
+                "links": [],
+                "interfaces": [],
+                "status": "pending",
+                "updated_at": now.strftime(TIME_FMT),
+                "inserted_at": now.strftime(TIME_FMT),
             },
             {
-                'id': '4567',
-                'start': start2.strftime(TIME_FMT),
-                'end': end2.strftime(TIME_FMT),
-                'switches': [
-                    '12:34:56:78:90:ab:cd:ef'
-                ],
-                'description': '',
-                'links': [],
-                'interfaces': [],
-                'status': 'pending',
-                'updated_at': now.strftime(TIME_FMT),
-                'inserted_at': now.strftime(TIME_FMT),
-            }
+                "id": "4567",
+                "start": start2.strftime(TIME_FMT),
+                "end": end2.strftime(TIME_FMT),
+                "switches": ["12:34:56:78:90:ab:cd:ef"],
+                "description": "",
+                "links": [],
+                "interfaces": [],
+                "status": "pending",
+                "updated_at": now.strftime(TIME_FMT),
+                "inserted_at": now.strftime(TIME_FMT),
+            },
         ]
 
-        url = f'{self.base_endpoint}'
+        url = f"{self.base_endpoint}"
         response = await self.api.get(url)
         current_data = response.json()
         assert response.status_code == 200
@@ -266,13 +402,12 @@ class TestMain:
     async def test_get_mw_case_3(self):
         """Test get non-existent id."""
         self.scheduler.get_maintenance.return_value = None
-        url = f'{self.base_endpoint}/2345'
+        url = f"{self.base_endpoint}/2345"
         response = await self.api.get(url)
         current_data = response.json()
         assert response.status_code == 404
-        assert current_data['description'] == \
-                         'Maintenance with id 2345 not found'
-        self.scheduler.get_maintenance.assert_called_once_with('2345')
+        assert current_data["description"] == "Maintenance with id 2345 not found"
+        self.scheduler.get_maintenance.assert_called_once_with("2345")
 
     async def test_get_mw_case_4(self):
         """Test get existent id."""
@@ -280,46 +415,41 @@ class TestMain:
         start2 = datetime.now(pytz.utc) + timedelta(hours=5)
         end2 = start2 + timedelta(hours=1, minutes=30)
         self.scheduler.get_maintenance.return_value = MW.construct(
-            id = '4567',
-            start = start2.replace(microsecond=0),
-            end = end2.replace(microsecond=0),
-            switches = [
-                '12:34:56:78:90:ab:cd:ef'
-            ],
-            updated_at = now.replace(microsecond=0),
-            inserted_at = now.replace(microsecond=0),
+            id="4567",
+            start=start2.replace(microsecond=0),
+            end=end2.replace(microsecond=0),
+            switches=["12:34:56:78:90:ab:cd:ef"],
+            updated_at=now.replace(microsecond=0),
+            inserted_at=now.replace(microsecond=0),
         )
         mw_dict = {
-            'id': '4567',
-            'start': start2.strftime(TIME_FMT),
-            'end': end2.strftime(TIME_FMT),
-            'switches': [
-                '12:34:56:78:90:ab:cd:ef'
-            ],
-            'description': '',
-            'links': [],
-            'interfaces': [],
-            'status': 'pending',
-            'updated_at': now.strftime(TIME_FMT),
-            'inserted_at': now.strftime(TIME_FMT),
+            "id": "4567",
+            "start": start2.strftime(TIME_FMT),
+            "end": end2.strftime(TIME_FMT),
+            "switches": ["12:34:56:78:90:ab:cd:ef"],
+            "description": "",
+            "links": [],
+            "interfaces": [],
+            "status": "pending",
+            "updated_at": now.strftime(TIME_FMT),
+            "inserted_at": now.strftime(TIME_FMT),
         }
-        url = f'{self.base_endpoint}/4567'
+        url = f"{self.base_endpoint}/4567"
         response = await self.api.get(url)
         current_data = response.json()
         assert response.status_code == 200
         assert current_data == mw_dict
-        self.scheduler.get_maintenance.assert_called_once_with('4567')
+        self.scheduler.get_maintenance.assert_called_once_with("4567")
 
     async def test_remove_mw_case_1(self):
         """Test remove non-existent id."""
         self.scheduler.get_maintenance.return_value = None
-        url = f'{self.base_endpoint}/2345'
+        url = f"{self.base_endpoint}/2345"
         response = await self.api.delete(url)
         current_data = response.json()
         assert response.status_code == 404
-        assert current_data['description'] == \
-                         'Maintenance with id 2345 not found'
-        self.scheduler.get_maintenance.assert_called_once_with('2345')
+        assert current_data["description"] == "Maintenance with id 2345 not found"
+        self.scheduler.get_maintenance.assert_called_once_with("2345")
         self.scheduler.remove.assert_not_called()
 
     async def test_remove_mw_case_2(self):
@@ -327,42 +457,41 @@ class TestMain:
         start1 = datetime.now(pytz.utc) + timedelta(hours=1)
         end1 = start1 + timedelta(hours=6)
         self.scheduler.get_maintenance.return_value = MW.construct(
-            id = '1234',
-            start = start1.replace(microsecond=0),
-            end = end1.replace(microsecond=0),
-            switches = [
-                '00:00:00:00:00:00:12:23'
-            ],
+            id="1234",
+            start=start1.replace(microsecond=0),
+            end=end1.replace(microsecond=0),
+            switches=["00:00:00:00:00:00:12:23"],
         )
-        url = f'{self.base_endpoint}/1234'
+        url = f"{self.base_endpoint}/1234"
         response = await self.api.delete(url)
         current_data = response.json()
         assert response.status_code == 200
-        assert current_data == {'response': 'Maintenance with id 1234 ' \
-                                                    'successfully removed'}
-        self.scheduler.get_maintenance.assert_called_once_with('1234')
-        self.scheduler.remove.assert_called_once_with('1234')
+        assert current_data == {
+            "response": "Maintenance with id 1234 " "successfully removed"
+        }
+        self.scheduler.get_maintenance.assert_called_once_with("1234")
+        self.scheduler.remove.assert_called_once_with("1234")
 
     async def test_remove_mw_case_3(self):
         """Test remove existent id."""
         start1 = datetime.now(pytz.utc) - timedelta(days=1)
         end1 = start1 + timedelta(hours=6)
         self.scheduler.get_maintenance.return_value = MW.construct(
-            id = '1234',
-            start = start1.replace(microsecond=0),
-            end = end1.replace(microsecond=0),
-            switches = [
-                '00:00:00:00:00:00:12:23'
-            ],
-            status = 'running',
+            id="1234",
+            start=start1.replace(microsecond=0),
+            end=end1.replace(microsecond=0),
+            switches=["00:00:00:00:00:00:12:23"],
+            status="running",
         )
-        url = f'{self.base_endpoint}/1234'
+        url = f"{self.base_endpoint}/1234"
         response = await self.api.delete(url)
         current_data = response.json()
         assert response.status_code == 400
-        assert current_data['description'] == \
-                         'Deleting a running maintenance is not allowed'
-        self.scheduler.get_maintenance.assert_called_once_with('1234')
+        assert (
+            current_data["description"]
+            == "Deleting a running maintenance is not allowed"
+        )
+        self.scheduler.get_maintenance.assert_called_once_with("1234")
         self.scheduler.remove.assert_not_called()
 
     async def test_update_mw_case_1(self, event_loop):
@@ -373,12 +502,11 @@ class TestMain:
         payload = {
             "start": start1.strftime(TIME_FMT),
         }
-        url = f'{self.base_endpoint}/2345'
+        url = f"{self.base_endpoint}/2345"
         response = await self.api.patch(url, json=payload)
         current_data = response.json()
         assert response.status_code == 404
-        assert current_data['description'] == \
-                         'Maintenance with id 2345 not found'
+        assert current_data["description"] == "Maintenance with id 2345 not found"
         self.scheduler.update.assert_not_called()
 
     async def test_update_mw_case_2(self, event_loop):
@@ -388,11 +516,11 @@ class TestMain:
         payload = {
             "start": start1.strftime(TIME_FMT),
         }
-        url = f'{self.base_endpoint}/1234'
+        url = f"{self.base_endpoint}/1234"
         response = await self.api.patch(url, json=payload)
         current_data = response.json()
         assert response.status_code == 400
-        assert "field required" in current_data['description']
+        assert "field required" in current_data["description"]
 
     async def test_update_mw_case_3(self, event_loop):
         """Test successful update."""
@@ -400,32 +528,27 @@ class TestMain:
         start1 = datetime.now(pytz.utc) + timedelta(days=1)
         end1 = start1 + timedelta(hours=6)
         self.scheduler.get_maintenance.return_value = MW.construct(
-            id = '1234',
-            start = start1.replace(microsecond=0),
-            end = end1.replace(microsecond=0),
-            switches = [
-                '00:00:00:00:00:00:12:23'
-            ],
+            id="1234",
+            start=start1.replace(microsecond=0),
+            end=end1.replace(microsecond=0),
+            switches=["00:00:00:00:00:00:12:23"],
         )
         start_new = datetime.now(pytz.utc) + timedelta(days=1, hours=3)
         payload = {
             "start": start_new.strftime(TIME_FMT),
         }
-        url = f'{self.base_endpoint}/1234'
+        url = f"{self.base_endpoint}/1234"
         response = await self.api.patch(url, json=payload)
         current_data = response.json()
-        assert current_data == \
-                         {'response': 'Maintenance 1234 updated'}
+        assert current_data == {"response": "Maintenance 1234 updated"}
         assert response.status_code == 200
-        self.scheduler.get_maintenance.assert_called_once_with('1234')
+        self.scheduler.get_maintenance.assert_called_once_with("1234")
         self.scheduler.update.assert_called_once_with(
             MW.construct(
-                id = '1234',
-                start = start_new.replace(microsecond=0),
-                end = end1.replace(microsecond=0),
-                switches = [
-                    '00:00:00:00:00:00:12:23'
-                ],
+                id="1234",
+                start=start_new.replace(microsecond=0),
+                end=end1.replace(microsecond=0),
+                switches=["00:00:00:00:00:00:12:23"],
             )
         )
 
@@ -435,24 +558,21 @@ class TestMain:
         start1 = datetime.now(pytz.utc) + timedelta(days=1)
         end1 = start1 + timedelta(hours=6)
         self.scheduler.get_maintenance.return_value = MW.construct(
-            id = '1234',
-            start = start1.replace(microsecond=0),
-            end = end1.replace(microsecond=0),
-            switches = [
-                '00:00:00:00:00:00:12:23'
-            ],
+            id="1234",
+            start=start1.replace(microsecond=0),
+            end=end1.replace(microsecond=0),
+            switches=["00:00:00:00:00:00:12:23"],
         )
         start_new = datetime.now(pytz.utc) - timedelta(days=1, hours=3)
         payload = {
             "start": start_new.strftime(TIME_FMT),
         }
-        url = f'{self.base_endpoint}/1234'
+        url = f"{self.base_endpoint}/1234"
         response = await self.api.patch(url, json=payload)
         current_data = response.json()
         assert response.status_code == 400
-        assert current_data['description'] == \
-                         'start: Start in the past not allowed'
-        self.scheduler.get_maintenance.assert_called_once_with('1234')
+        assert current_data["description"] == "start: Start in the past not allowed"
+        self.scheduler.get_maintenance.assert_called_once_with("1234")
         self.scheduler.update.assert_not_called()
 
     async def test_update_mw_case_5(self, event_loop):
@@ -461,26 +581,23 @@ class TestMain:
         start1 = datetime.now(pytz.utc) + timedelta(days=1)
         end1 = start1 + timedelta(hours=6)
         self.scheduler.get_maintenance.return_value = MW.construct(
-            id = '1234',
-            start = start1.replace(microsecond=0),
-            end = end1.replace(microsecond=0),
-            switches = [
-                '00:00:00:00:00:00:12:23'
-            ],
+            id="1234",
+            start=start1.replace(microsecond=0),
+            end=end1.replace(microsecond=0),
+            switches=["00:00:00:00:00:00:12:23"],
         )
         start_new = datetime.now(pytz.utc) + timedelta(days=1, hours=3)
         end_new = start_new - timedelta(hours=5)
         payload = {
             "start": start_new.strftime(TIME_FMT),
-            "end": end_new.strftime(TIME_FMT)
+            "end": end_new.strftime(TIME_FMT),
         }
-        url = f'{self.base_endpoint}/1234'
+        url = f"{self.base_endpoint}/1234"
         response = await self.api.patch(url, json=payload)
         current_data = response.json()
         assert response.status_code == 400
-        assert current_data['description'] == \
-                         'end: End before start not allowed'
-        self.scheduler.get_maintenance.assert_called_once_with('1234')
+        assert current_data["description"] == "end: End before start not allowed"
+        self.scheduler.get_maintenance.assert_called_once_with("1234")
         self.scheduler.update.assert_not_called()
 
     async def test_update_mw_case_6(self, event_loop):
@@ -489,28 +606,28 @@ class TestMain:
         start1 = datetime.now(pytz.utc) + timedelta(days=1)
         end1 = start1 + timedelta(hours=6)
         self.scheduler.get_maintenance.return_value = MW.construct(
-            id = '1234',
-            start = start1.replace(microsecond=0),
-            end = end1.replace(microsecond=0),
-            switches = [
-                '00:00:00:00:00:00:12:23'
-            ],
+            id="1234",
+            start=start1.replace(microsecond=0),
+            end=end1.replace(microsecond=0),
+            switches=["00:00:00:00:00:00:12:23"],
         )
         start_new = datetime.now(pytz.utc) + timedelta(days=1, hours=3)
         payload = {
             "start": start_new.strftime(TIME_FMT),
             "switches": [],
-            'interfaces': [],
-            'links': [],
+            "interfaces": [],
+            "links": [],
         }
 
-        url = f'{self.base_endpoint}/1234'
+        url = f"{self.base_endpoint}/1234"
         response = await self.api.patch(url, json=payload)
         current_data = response.json()
         assert response.status_code == 400
-        assert current_data['description'] == \
-                         '__root__: At least one item must be provided'
-        self.scheduler.get_maintenance.assert_called_once_with('1234')
+        assert (
+            current_data["description"]
+            == "__root__: At least one item must be provided"
+        )
+        self.scheduler.get_maintenance.assert_called_once_with("1234")
         self.scheduler.update.assert_not_called()
 
     async def test_update_mw_case_7(self, event_loop):
@@ -519,99 +636,94 @@ class TestMain:
         start1 = datetime.now(pytz.utc) + timedelta(days=1)
         end1 = start1 + timedelta(hours=6)
         self.scheduler.get_maintenance.return_value = MW.construct(
-            id = '1234',
-            start = start1.replace(microsecond=0),
-            end = end1.replace(microsecond=0),
-            switches = [
-                '00:00:00:00:00:00:12:23'
-            ],
+            id="1234",
+            start=start1.replace(microsecond=0),
+            end=end1.replace(microsecond=0),
+            switches=["00:00:00:00:00:00:12:23"],
         )
         payload = {
-            'status': 'running',
+            "status": "running",
         }
-        url = f'{self.base_endpoint}/1234'
+        url = f"{self.base_endpoint}/1234"
         response = await self.api.patch(url, json=payload)
         current_data = response.json()
         assert response.status_code == 400
-        assert current_data['description'] == \
-                         'Updating a maintenance status is not allowed'
+        assert (
+            current_data["description"]
+            == "Updating a maintenance status is not allowed"
+        )
         self.scheduler.update.assert_not_called()
 
     async def test_end_mw_case_1(self):
         """Test method that finishes the maintenance now."""
         self.scheduler.get_maintenance.return_value = None
-        url = f'{self.base_endpoint}/2345/end'
+        url = f"{self.base_endpoint}/2345/end"
         response = await self.api.patch(url)
         current_data = response.json()
         assert response.status_code == 404
-        assert current_data['description'] == \
-                         'Maintenance with id 2345 not found'
+        assert current_data["description"] == "Maintenance with id 2345 not found"
 
     async def test_end_mw_case_2(self):
         """Test method that finishes the maintenance now."""
         start1 = datetime.now(pytz.utc) - timedelta(hours=1)
         end1 = start1 + timedelta(hours=6)
         self.scheduler.get_maintenance.return_value = MW.construct(
-            id = '1234',
-            start = start1.replace(microsecond=0),
-            end = end1.replace(microsecond=0),
-            switches = [
-                '00:00:00:00:00:00:12:23'
-            ],
-            status = 'running',
+            id="1234",
+            start=start1.replace(microsecond=0),
+            end=end1.replace(microsecond=0),
+            switches=["00:00:00:00:00:00:12:23"],
+            status="running",
         )
-        url = f'{self.base_endpoint}/1234/end'
+        url = f"{self.base_endpoint}/1234/end"
         response = await self.api.patch(url)
         current_data = response.json()
-        self.scheduler.get.asssert_called_once_with('1234')
-        self.scheduler.end_maintenance_early.assert_called_once_with('1234')
+        self.scheduler.get.asssert_called_once_with("1234")
+        self.scheduler.end_maintenance_early.assert_called_once_with("1234")
         assert response.status_code == 200
-        assert current_data == \
-                         {'response': 'Maintenance window 1234 finished'}
+        assert current_data == {"response": "Maintenance window 1234 finished"}
 
     async def test_end_mw_case_3(self):
         """Test method that finishes the maintenance now."""
         start1 = datetime.now(pytz.utc) + timedelta(hours=1)
         end1 = start1 + timedelta(hours=6)
         self.scheduler.get_maintenance.return_value = MW.construct(
-            id = '1234',
-            start = start1.replace(microsecond=0),
-            end = end1.replace(microsecond=0),
-            switches = [
-                '00:00:00:00:00:00:12:23'
-            ],
-            status = 'pending',
+            id="1234",
+            start=start1.replace(microsecond=0),
+            end=end1.replace(microsecond=0),
+            switches=["00:00:00:00:00:00:12:23"],
+            status="pending",
         )
-        url = f'{self.base_endpoint}/1234/end'
+        url = f"{self.base_endpoint}/1234/end"
         response = await self.api.patch(url)
         current_data = response.json()
-        self.scheduler.get.asssert_called_once_with('1234')
+        self.scheduler.get.asssert_called_once_with("1234")
         self.scheduler.end_maintenance_early.assert_not_called()
         assert response.status_code == 400
-        assert current_data['description'] == \
-                         'Maintenance window 1234 has not yet started'
+        assert (
+            current_data["description"] == "Maintenance window 1234 has not yet started"
+        )
 
     async def test_end_mw_case_4(self):
         """Test method that finishes the maintenance now."""
         start1 = datetime.now(pytz.utc) - timedelta(hours=5)
         end1 = start1 + timedelta(hours=4)
         self.scheduler.get_maintenance.return_value = MW.construct(
-            id = '1234',
-            start = start1.replace(microsecond=0),
-            end = end1.replace(microsecond=0),
-            switches = [
-                '00:00:00:00:00:00:12:23'
-            ],
-            status = 'finished',
+            id="1234",
+            start=start1.replace(microsecond=0),
+            end=end1.replace(microsecond=0),
+            switches=["00:00:00:00:00:00:12:23"],
+            status="finished",
         )
-        url = f'{self.base_endpoint}/1234/end'
+        url = f"{self.base_endpoint}/1234/end"
         response = await self.api.patch(url)
         current_data = response.json()
-        self.scheduler.get.asssert_called_once_with('1234')
+        self.scheduler.get.asssert_called_once_with("1234")
         self.scheduler.end_maintenance_early.assert_not_called()
         assert response.status_code == 400
-        assert current_data['description'] == \
-                         'Maintenance window 1234 has already finished'
+        assert (
+            current_data["description"]
+            == "Maintenance window 1234 has already finished"
+        )
 
     async def test_extend_case_1(self, event_loop):
         """Test successful extension."""
@@ -619,67 +731,57 @@ class TestMain:
         start1 = datetime.now(pytz.utc) - timedelta(hours=3)
         end1 = start1 + timedelta(hours=4)
         self.scheduler.get_maintenance.return_value = MW.construct(
-            id = '1234',
-            start = start1.replace(microsecond=0),
-            end = end1.replace(microsecond=0),
-            switches = [
-                '00:00:00:00:00:00:12:23'
-            ],
-            status = 'running',
+            id="1234",
+            start=start1.replace(microsecond=0),
+            end=end1.replace(microsecond=0),
+            switches=["00:00:00:00:00:00:12:23"],
+            status="running",
         )
-        url = f'{self.base_endpoint}/1234/extend'
-        payload = {
-            'minutes': 45
-        }
+        url = f"{self.base_endpoint}/1234/extend"
+        payload = {"minutes": 45}
         response = await self.api.patch(url, json=payload)
         assert response.status_code == 200
-        self.scheduler.get_maintenance.called_with('1234')
-        self.scheduler.update.assert_called_with(MW.construct(
-            id = '1234',
-            start = start1.replace(microsecond=0),
-            end = end1.replace(microsecond=0) + timedelta(minutes=45),
-            switches = [
-                '00:00:00:00:00:00:12:23'
-            ],
-            status = 'running',
-        ))
+        self.scheduler.get_maintenance.called_with("1234")
+        self.scheduler.update.assert_called_with(
+            MW.construct(
+                id="1234",
+                start=start1.replace(microsecond=0),
+                end=end1.replace(microsecond=0) + timedelta(minutes=45),
+                switches=["00:00:00:00:00:00:12:23"],
+                status="running",
+            )
+        )
 
     async def test_extend_case_2(self, event_loop):
         """Test no payload error."""
         self.napp.controller.loop = event_loop
-        url = f'{self.base_endpoint}/1234/extend'
+        url = f"{self.base_endpoint}/1234/extend"
         response = await self.api.patch(url)
         assert response.status_code == 400
         current_data = response.json()
-        assert 'Invalid json' in current_data['description']
+        assert "Invalid json" in current_data["description"]
         self.scheduler.update.assert_not_called()
 
     async def test_extend_case_3(self, event_loop):
         """Test payload without minutes."""
         self.napp.controller.loop = event_loop
-        url = f'{self.base_endpoint}/1234/extend'
-        payload = {
-            'seconds': 240
-        }
+        url = f"{self.base_endpoint}/1234/extend"
+        payload = {"seconds": 240}
         response = await self.api.patch(url, json=payload)
         assert response.status_code == 400
         current_data = response.json()
-        assert current_data['description'] == \
-                         'Minutes of extension must be sent'
+        assert current_data["description"] == "Minutes of extension must be sent"
         self.scheduler.update.assert_not_called()
 
     async def test_extend_case_4(self, event_loop):
         """Test no integer extension minutes."""
         self.napp.controller.loop = event_loop
-        url = f'{self.base_endpoint}/1234/extend'
-        payload = {
-            'minutes': '240'
-        }
+        url = f"{self.base_endpoint}/1234/extend"
+        payload = {"minutes": "240"}
         response = await self.api.patch(url, json=payload)
         assert response.status_code == 400
         current_data = response.json()
-        assert current_data['description'] == \
-                         'Minutes of extension must be integer'
+        assert current_data["description"] == "Minutes of extension must be integer"
         self.scheduler.update.assert_not_called()
 
     async def test_extend_case_5(self, event_loop):
@@ -688,24 +790,21 @@ class TestMain:
         start1 = datetime.now(pytz.utc) + timedelta(hours=3)
         end1 = start1 + timedelta(hours=4)
         self.scheduler.get_maintenance.return_value = MW.construct(
-            id = '1234',
-            start = start1.replace(microsecond=0),
-            end = end1.replace(microsecond=0),
-            switches = [
-                '00:00:00:00:00:00:12:23'
-            ],
-            status = 'pending',
+            id="1234",
+            start=start1.replace(microsecond=0),
+            end=end1.replace(microsecond=0),
+            switches=["00:00:00:00:00:00:12:23"],
+            status="pending",
         )
-        url = f'{self.base_endpoint}/1234/extend'
-        payload = {
-            'minutes': 240
-        }
+        url = f"{self.base_endpoint}/1234/extend"
+        payload = {"minutes": 240}
         response = await self.api.patch(url, json=payload)
         assert response.status_code == 400
         current_data = response.json()
-        assert current_data['description'] == \
-                         'Maintenance window 1234 has not yet started'
-        self.scheduler.get_maintenance.assert_called_once_with('1234')
+        assert (
+            current_data["description"] == "Maintenance window 1234 has not yet started"
+        )
+        self.scheduler.get_maintenance.assert_called_once_with("1234")
         self.scheduler.update.assert_not_called()
 
     async def test_extend_case_6(self, event_loop):
@@ -714,38 +813,33 @@ class TestMain:
         start1 = datetime.now(pytz.utc) - timedelta(hours=3)
         end1 = start1 + timedelta(hours=2)
         self.scheduler.get_maintenance.return_value = MW.construct(
-            id = '1234',
-            start = start1.replace(microsecond=0),
-            end = end1.replace(microsecond=0),
-            switches = [
-                '00:00:00:00:00:00:12:23'
-            ],
-            status = 'finished',
+            id="1234",
+            start=start1.replace(microsecond=0),
+            end=end1.replace(microsecond=0),
+            switches=["00:00:00:00:00:00:12:23"],
+            status="finished",
         )
-        url = f'{self.base_endpoint}/1234/extend'
-        payload = {
-            'minutes': 240
-        }
+        url = f"{self.base_endpoint}/1234/extend"
+        payload = {"minutes": 240}
         response = await self.api.patch(url, json=payload)
         assert response.status_code == 400
         current_data = response.json()
-        assert current_data['description'] == \
-                         'Maintenance window 1234 has already finished'
-        self.scheduler.get_maintenance.assert_called_once_with('1234')
+        assert (
+            current_data["description"]
+            == "Maintenance window 1234 has already finished"
+        )
+        self.scheduler.get_maintenance.assert_called_once_with("1234")
         self.scheduler.update.assert_not_called()
 
     async def test_extend_case_7(self, event_loop):
         """Test no maintenace found."""
         self.napp.controller.loop = event_loop
         self.scheduler.get_maintenance.return_value = None
-        url = f'{self.base_endpoint}/1235/extend'
-        payload = {
-            'minutes': 240
-        }
+        url = f"{self.base_endpoint}/1235/extend"
+        payload = {"minutes": 240}
         response = await self.api.patch(url, json=payload)
         assert response.status_code == 404
         current_data = response.json()
-        assert current_data['description'] == \
-                         'Maintenance with id 1235 not found'
-        self.scheduler.get_maintenance.assert_called_once_with('1235')
+        assert current_data["description"] == "Maintenance with id 1235 not found"
+        self.scheduler.get_maintenance.assert_called_once_with("1235")
         self.scheduler.update.assert_not_called()